[ALPS04267334] FPSGO: XGF v2 fix coverity of dep recycle
 FPSGO: XGF v2 fix coverity of dep recycle

MTK-Commit-Id: fb39a322b2ca32b4cff189f8cb5ded3952b92bcc

CR-Id: ALPS04267334
Feature: FPSGO

+++ drivers/misc/mediatek/performance/fpsgo_v2/fbt/src/xgf.c
@@ -458,7 +458,7 @@ static inline void xgf_dep_recycle(struct rb_root *root,
 
 	diff = (long long)now_ts - (long long)last_check2recycle_dep_ts;
 
-	if (diff < 0LL || diff < (NSEC_PER_SEC << shift))
+	if (diff < 0LL || diff < ((long long)NSEC_PER_SEC << shift))
 		goto out;
 
 	n = rb_first(root);
